Move some pytest fixtures out of `conftest.py`
