+++ tests/desmume/test_shop_items.py
@@ -1,4 +1,45 @@
+import os
+from pathlib import Path
+
+from ndspy.rom import NintendoDSRom
+import pytest
+
+from patcher.location_types import IslandShopLocation
+from patcher.location_types.island_shop import GD_MODELS
+
 from .desmume_utils import DesmumeEmulator, get_current_rupee_count, start_first_file
+
+
+@pytest.fixture(
+    params=[val for val in GD_MODELS.keys() if GD_MODELS[val]],
+    ids=[f'{hex(key)}-{val}' for key, val in GD_MODELS.items() if val],
+)
+def island_shop_test_emu(tmp_path: Path, desmume_emulator: DesmumeEmulator, request):
+    test_name = (
+        os.environ['PYTEST_CURRENT_TEST']
+        .split(':')[-1]
+        .split(' ')[0]
+        .replace('[', '_')
+        .replace(']', '_')
+    )
+    rom_path = str(tmp_path / f'{test_name}.nds')
+
+    IslandShopLocation.ROM = NintendoDSRom.fromFile(rom_path)
+
+    locations = [
+        IslandShopLocation(31, 0x217ECB4 - 0x217BCE0),  # shield in mercay shop
+        IslandShopLocation(31, 0x217EC68 - 0x217BCE0),  # power gem in mercay shop
+        IslandShopLocation(31, 0x217EC34 - 0x217BCE0),  # treasure item in mercay shop
+    ]
+
+    for location in locations:
+        location.set_location(request.param)
+
+    IslandShopLocation.ROM.saveToFile(rom_path)
+
+    desmume_emulator.open_rom(rom_path)
+
+    return desmume_emulator
 
 
 def test_custom_shop_items(island_shop_test_emu: DesmumeEmulator):
